# resources/openapi/openapi-core.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Baobab CMS API",
        "version": "1.0.0",
        "description": "REST v1 — registre des Content Types (spec 08 §2). Généré depuis les blueprints actifs, jamais édité à la main."
    },
    "paths": {
        "/users": {
            "get": {
                "summary": "Liste des utilisateurs",
                "tags": [
                    "users"
                ],
                "description": "Sous baobab.users.manage ou baobab.users.impersonate (spec 05 §7).",
                "parameters": [
                    {
                        "name": "filter[role]",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[pending]",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "filter[deactivated]",
                        "in": "query",
                        "required": false,
                        "description": "true : uniquement les comptes désactivés ; false : uniquement les autres.",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "maximum": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    }
                }
            },
            "post": {
                "summary": "Inviter un utilisateur",
                "tags": [
                    "users"
                ],
                "description": "Sous baobab.users.manage. Le rôle doit être de niveau strictement inférieur à celui de l'appelant ; l'invité reçoit un lien valable 24 heures (spec 05 décision 5).",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "name",
                                    "email",
                                    "role"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "role": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Invité",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                }
            }
        },
        "/users/{user}": {
            "get": {
                "summary": "Afficher un utilisateur",
                "tags": [
                    "users"
                ],
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            },
            "patch": {
                "summary": "Modifier le nom ou l'e-mail d'un utilisateur (validé par e-mail)",
                "description": "Le titulaire ou un admin qui le domine (baobab.users.manage). Un compte dont l'invitation est en attente n'a pas d'e-mail modifiable : annuler l'invitation, puis inviter la bonne adresse.",
                "tags": [
                    "users"
                ],
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255
                                    },
                                    "lost_mailbox": {
                                        "type": "boolean",
                                        "description": "Exception « boîte perdue » : réservée à un admin qui domine le compte, ne change que l'e-mail. Exige admin_password et justification ; la demande part à la nouvelle adresse."
                                    },
                                    "admin_password": {
                                        "type": "string",
                                        "description": "Mot de passe de l'admin qui fait la demande (« boîte perdue » seulement)."
                                    },
                                    "justification": {
                                        "type": "string",
                                        "maxLength": 1000,
                                        "description": "Obligatoire pour « boîte perdue » ; conservée dans l'audit."
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Demande enregistrée. Rien n'est modifié tant que le lien envoyé par e-mail n'est pas confirmé : à l'adresse actuelle du compte (puis à la nouvelle pour un e-mail), ou à la nouvelle adresse seule pour « boîte perdue ».",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "status": {
                                                    "type": "string",
                                                    "enum": [
                                                        "pending"
                                                    ]
                                                },
                                                "stage": {
                                                    "type": "string",
                                                    "enum": [
                                                        "verify",
                                                        "confirm"
                                                    ],
                                                    "description": "verify : en attente de l'adresse actuelle ; confirm : en attente de la nouvelle."
                                                },
                                                "new_name": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "new_email": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "lost_mailbox": {
                                                    "type": "boolean"
                                                },
                                                "expires_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                }
            }
        },
...
        "/me": {
            "patch": {
                "summary": "Modifier son propre nom ou e-mail (validé par e-mail)",
                "description": "Le compte authentifié. Mêmes règles que PATCH /users/{user}, sans l'exception « boîte perdue ».",
                "tags": [
                    "users"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Demande enregistrée. Rien n'est modifié tant que le lien envoyé par e-mail n'est pas confirmé : à l'adresse actuelle du compte (puis à la nouvelle pour un e-mail), ou à la nouvelle adresse seule pour « boîte perdue ».",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "status": {
                                                    "type": "string",
                                                    "enum": [
                                                        "pending"
                                                    ]
                                                },
                                                "stage": {
                                                    "type": "string",
                                                    "enum": [
                                                        "verify",
                                                        "confirm"
                                                    ],
                                                    "description": "verify : en attente de l'adresse actuelle ; confirm : en attente de la nouvelle."
                                                },
                                                "new_name": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "new_email": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "lost_mailbox": {
                                                    "type": "boolean"
                                                },
                                                "expires_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "Jeton personnel Sanctum (spec 08 §4.1, admin/account/api-tokens)."
            },
            "sessionCookie": {
                "type": "apiKey",
                "in": "cookie",
                "name": "baobab_session",
                "description": "Session du guard baobab (utilisée par l'admin)."
            }
        },
        "schemas": {
            "ProblemDetails": {
                "type": "object",
                "description": "Erreur RFC 9457 (spec 08 §2.2, Baobab\\Api\\Support\\ProblemDetailsRenderer).",
                "required": [
                    "type",
                    "title",
                    "status"
                ],
                "properties": {
                    "type": {
                        "type": "string",
                        "format": "uri"
                    },
                    "title": {
                        "type": "string"
                    },
                    "status": {
                        "type": "integer"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "User": {
                "type": "object",
                "description": "Utilisateur (spec 05 §7) — jamais de secret.",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "level": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "level": {
                        "type": "integer"
                    },
                    "invitation_pending": {
                        "type": "boolean"
                    },
                    "invited_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    },
                    "deactivated": {
                        "type": "boolean",
                        "description": "Compte désactivé par un administrateur."
                    },
                    "deactivated_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    },
                    "deactivation_reason": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "Motif facultatif de la désactivation."
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    }
                }
            }
        },
        "responses": {
            "Unauthenticated": {
                "description": "Authentification requise.",
                "content": {
                    "application/problem+json": {
                        "schema": {
                            "$ref": "#/components/schemas/ProblemDetails"
                        }
                    }
                }
            },
            "Forbidden": {
                "description": "Permission manquante.",
                "content": {
                    "application/problem+json": {
                        "schema": {
                            "$ref": "#/components/schemas/ProblemDetails"
                        }
                    }
                }
            },
            "NotFound": {
                "description": "Type ou entrée introuvable.",
                "content": {
                    "application/problem+json": {
                        "schema": {
                            "$ref": "#/components/schemas/ProblemDetails"
                        }
                    }
                }
            },
...
            "ValidationError": {
                "description": "Données invalides — `errors` porte le détail par champ.",
                "content": {
                    "application/problem+json": {
                        "schema": {
                            "$ref": "#/components/schemas/ProblemDetails"
                        }
                    }
                }
            },
...
        }
    }
}